+++ src/main/java/net/geforcemods/securitycraft/api/IOwnable.java
@@ -1,11 +1,8 @@
 package net.geforcemods.securitycraft.api;
 
 import net.geforcemods.securitycraft.SCContent;
-import net.geforcemods.securitycraft.items.ItemUniversalOwnerChanger;
 import net.geforcemods.securitycraft.util.PlayerUtils;
 import net.geforcemods.securitycraft.util.Utils;
-import net.minecraft.block.BlockDoor;
-import net.minecraft.block.BlockDoor.EnumDoorHalf;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.tileentity.TileEntity;
@@ -48,8 +45,7 @@ public interface IOwnable {
 	}
 
 	/**
-	 * Executes actions after the owner has been changed, for example making sure the owner of both halves of SecurityCraft's
-	 * doors get changed, and marks IOwnables that need validation as invalidated
+	 * Executes after the owner has been changed and invalidates this if it needs validation
 	 *
 	 * @param world The current world
 	 * @param state The IOwnable's state
@@ -57,9 +53,6 @@ public interface IOwnable {
 	 * @param player The player that changed the owner of the IOwnable
 	 */
 	default void onOwnerChanged(IBlockState state, World world, BlockPos pos, EntityPlayer player) {
-		if (state.getPropertyKeys().contains(BlockDoor.HALF))
-			ItemUniversalOwnerChanger.tryUpdateBlock(world, state.getValue(BlockDoor.HALF) == EnumDoorHalf.UPPER ? pos.down() : pos.up(), getOwner());
-
 		if (needsValidation()) {
 			getOwner().setValidated(false);
 			PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.universalOwnerChanger.getTranslationKey()), Utils.localize("messages.securitycraft:universalOwnerChanger.ownerInvalidated"), TextFormatting.GREEN);