+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fix: Reinforced pressure plates cannot open reinforced doors/trapdoors/etc.
 - Fix: Password-protected Furnace is not shown as a recipe catalyst for furnace recipes in JEI
 - Fix: Light level of Password-protected Furnace is too high compared to the vanilla furnace
+- Fix: Changing the owner of the Inventory Scanner or doors does not change the owner of the other block/half
 
 --------------------------Changelog for v1.9.0.2 of SecurityCraft--------------------------